requirements.txt
# Core computational dependencies
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
cadquery==2.6.1
cadquery-ocp==7.8.1.1.post1
cadquery_vtk==9.3.1
casadi==3.7.2
choreographer==1.2.1
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
Cython==0.29.37
dataclasses-json==0.6.7
deap==1.4.3
Deprecated==1.3.1
et_xmlfile==2.0.0
ezdxf==1.4.3
fastapi==0.121.2
fonttools==4.60.1
freetype-py==2.5.1
frozenlist==1.8.0
h11==0.16.0
httptools==0.7.1
idna==3.11
ImageIO==2.37.2
iniconfig==2.3.0
kaleido==1.2.0
kiwisolver==1.4.9
logistro==2.0.1
marshmallow==3.26.1
matplotlib==3.10.7
more-itertools==10.8.0
msgpack==1.1.2
multidict==6.7.0
multimethod==1.12
mypy_extensions==1.1.0
narwhals==2.12.0
networkx==3.5
nlopt==2.9.1
numpy==1.26.4
openpyxl==3.1.5
orjson==3.11.4
packaging==25.0
pandas==2.3.3
path==17.1.1
pillow==12.0.0
pip==25.2
plotly==6.5.0
pluggy==1.6.0
propcache==0.4.1
py3dbp==1.1.2
pydantic==2.12.4
pydantic_core==2.41.5
pyembree==0.1.12
pyglet==2.1.11
pygltflib==1.16.5
Pygments==2.19.2
PyOpenGL==3.1.0
pyparsing==3.2.5
pyrender==0.1.45
pytest==9.0.1
pytest-timeout==2.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
rectpack==0.2.2
rtree==1.4.1
scipy==1.16.3
setuptools==60.10.0
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
starlette==0.49.3
tqdm==4.67.1
trame==3.12.0
trame-client==3.11.2
trame-common==1.0.1
trame-server==3.8.1
trame-vtk==2.10.0
trimesh==3.23.5
typing_extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typish==1.9.3
tzdata==2025.2
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
wheel==0.37.1
wrapt==2.0.1
wslink==2.5.0
yarl==1.22.0

# Django and REST framework dependencies
Django==5.1
djangorestframework==3.15.2
psycopg2-binary==2.9.10
django-cors-headers==4.6.0
django-filter==24.3
drf-spectacular==0.28.0
python-decouple==3.8

# Optional: Celery for async tasks (future enhancement)
# celery[redis]==5.4.0
# redis==5.2.0

# Optional: Production server
# gunicorn==23.0.0
